backend/requirements.txt
# Web Framework
fastapi==0.115.6
uvicorn[standard]==0.32.1

# Database
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
alembic==1.14.0

# Reddit API
praw==7.8.1

# Data Processing
pandas==2.2.3
numpy==1.26.4

# Async Support
asyncio-mqtt==0.16.2
asyncpg==0.30.0

# Configuration
python-dotenv==1.0.1
pydantic==2.10.5

# Analytics & ML
scikit-learn==1.5.2
textblob==0.18.0
wordcloud==1.9.3

# Export Formats
pyarrow==18.1.0

# Testing
pytest==8.3.4
httpx==0.28.1

# Utilities
rich==13.9.4
click==8.1.8
tqdm==4.67.1